common: update Travis to Ubuntu 18.04.3 LTS Bionic
We need Xenial or Bionic to enable ppc64le builds[1], so lets go to
the more recent one. This forces us to deal with the previous issue
with bind mount mentioned in:

    commit: fc2257f5233677b6b0aa39acecaf63ec33fc1280

Instead of use user namespace remap this can be fixed with a chown.
The chown can be done from outside or inside docker, but from inside
we don't need to know the UID/GID of the docker user nor try to find
it out.

[1] https://docs.travis-ci.com/user/reference/overview/
    #virtualisation-environment-vs-operating-system

Signed-off-by: Raoni Fassina Firmino <raoni@linux.ibm.com>