common: add ppc64le Travis build
Add support to ppc64le builds in Travis. This duplicates the number of
jobs that runs from Travis. All jobs from the matrix are supported in
ppc64le.

Support for ppc64le builds runs only on Xenial & Bionic and only from
inside a LXD container[1]. There is a limitation in running Docker
from inside LXD, the ones relevant for us are: '--privileged' and
'-ti'. I am not sure we really need the later and the former is only
needed to create a tmpfs /tmp. The tmpfs can be done from outside docker
and just exported to it with a bind-mount or directly using the
specific '--tmpfs' docker-run option. I choose the later. For the
output/tty redirection, for now I am disabling it for ppc64le builds
(part of it is already disabled for GHA), and if it is not really
needed we can remove it all together.

Some parts of docker image build and package generation had x86_84
hardcoded and needed to be fixed.

Apparently, for some reason, docker pull/run will download and run an
docker image of different architecture if it does not find a matching
one. Without warning or anything. And because PMDK only have and amd64
image on docker hub we add this workaround to detect if an docker
image is of an supported architecture for the host.

[1] https://docs.travis-ci.com/user/reference/overview/
    #virtualisation-environment-vs-operating-system

Signed-off-by: Raoni Fassina Firmino <raoni@linux.ibm.com>