common: add support for per arch variables on travis
Travis don't have a simple way to define per arch "global" variables.
So or option was to duplicate all jobs for each architecture or hack a
way to define per architecture variables. I choose the later in a
generic way to support any number of variables with an arch-prefix.

With this support, on travis we can define variables like so:

    env:
      global:
        - VALGRIND=1
        - ppc64le__VALGRIND=0

And the resulting VALGRIND variable will be exported with different
values depending $TRAVIS_CPU_ARCH [1][2] that the jobs is running on.
It will export the same env to all subsequent command in docker
`before_install` and `script` stages (and all following stages)

The sintax for this variables is to add a prefix in the form
${TRAVIS_CPU_ARCH}__ to any variable.

This is needed to add two ppc64le specific variables:

- BLACKLIST_FILE, to filter know failing tests from the CI build and
  test.
- VALGRIND, to force disable valgrind in all ppc64le testes. For now
  pmem/valgrind  does not support ppc64le but it successful build and
  install during ppc64le docker image generation, so the RUNTESTS
  don't skip them. An other approach would be to conditionally install
  pmem/valgrind, but I choose the more explicitly one.

[1] https://docs.travis-ci.com/user/multi-cpu-architectures/
    #identifying-cpu-architecture-of-build-jobs
[2] https://docs.travis-ci.com/user/environment-variables/
    #default-environment-variables

Signed-off-by: Raoni Fassina Firmino <raoni@linux.ibm.com>