common: rewrite ppc64le.blacklist with current data
This list does not include pmemcheck failling tests due to pmemcheck
being broken on ppc64le.

This list was generated with the falling tests on travis with ppc64le
jobs. This is done using the distro's default valgrind, that don't
support pmemcheck and because of it don't run tests that need it, but
runs more tests than forcing VALGRIND=0.

It also includes obj_defrag_advanced, that in fact works but runs
slower on ppc64le and so keeps timing out with the default 3m config.

Signed-off-by: Raoni Fassina Firmino <raoni@linux.ibm.com>