common: explicitly enable disabled jobs for ppc64le only
The disabled amd64 jobs were moved to be ppc64le only. this
configuration exploit a defined behavior of matrix.include:

> The jobs which are explicitly included inherit the first value of
> the expansion keys defined [1]

In this case the only expansion key with more than one entry is the
"arch" key, and that is why ppc64le moved to be de first of the list,
to avoid have to define "arch:" for each entry in matrix.include and
have it unnecessarily verbose and less readable.

[1] https://docs.travis-ci.com/user/build-matrix/
    #explicitly-included-jobs-inherit-the-first-value-in-the-array

Signed-off-by: Raoni Fassina Firmino <raoni@linux.ibm.com>